requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-dotenv
faker
